+++ HISTORY.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-1.0 (unreleased)
+1.0 (2014-06-18)
 ****************
 
 - Added themes!